requirements.txt
aiohttp==3.6.2
discord.py==1.4.1
tinydb==4.1.1
